Move LinkArrowIcon to the UI folder
