+++ package.json
@@ -44,7 +44,7 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.5",
     "prettier": "3.2.5",
-    "typescript": "^5.2.2",
+    "typescript": "^5.7.3",
     "vite": "^5.0.8"
   }
 }
